Move web navigation delegate methods into TabObserver
