+++ ios/brave-ios/Sources/Brave/Frontend/Browser/BrowserViewController/BVC+TabManagerDelegate.swift
@@ -192,7 +192,6 @@ extension BrowserViewController: TabManagerDelegate {
     tab.tabDelegate = self
     tab.addObserver(self)
     tab.addPolicyDecider(self)
-    tab.navigationDelegate = self
     tab.webDelegate = self
     tab.downloadDelegate = self
     tab.certStore = profile.certStore
